feat(example): ✨ add JSON output switch
