+++ DomainDetective.CLI/DomainDetective.CLI.csproj
@@ -9,8 +9,4 @@
     <InvariantGlobalization>true</InvariantGlobalization>
   </PropertyGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\DomainDetective\DomainDetective.csproj" />
-  </ItemGroup>
-
 </Project>
